+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 * [HexaPDF::Layout::Style] property `last_line_gap` for controlling the spacing
   after the last line of text
 * [HexaPDF::Layout::Box#draw_content] for use by subclasses
+* [HexaPDF::Type::Form#width] and [HexaPDF::Type::Form#height] for compatibility
+  with [HexaPDF::Type::Image]
 
 ### Changed